refactor: token 초기화로직 분리
