+++ src/test/java/com/camping/admin/helper/CommonContext.java
@@ -12,6 +12,9 @@ import static io.restassured.RestAssured.given;
 import static io.restassured.http.ContentType.JSON;
 
 public class CommonContext {
+
+    public static String adminToken;
+
     public static RequestSpecification getRequestSpec() {
         return new RequestSpecBuilder()
                 .setBaseUri("http://localhost")
@@ -20,18 +23,5 @@ public class CommonContext {
                 .setContentType(JSON)
                 .log(LogDetail.ALL)
                 .build();
-    }
-
-    public static String getAdminToken() {
-        return given()
-                .contentType(JSON)
-                .body(Map.of("username", "admin", "password", "admin123"))
-                .when()
-                .post("/auth/login")
-                .then()
-                .log().all()
-                .extract()
-                .jsonPath()
-                .getString("accessToken");
     }
 }
